feat: Add find-jobs step to build-dev-container.yaml
This commit adds a new step called "find-jobs" to the build-dev-container.yaml workflow. The step uses the "philips-labs/list-folders-action@v1" action to list folders in the "./templates" directory and outputs the result as "folders". The step runs in a container based on Debian Buster Slim.

The commit also includes commented out code for the "matrix" and "build" steps, which are currently disabled.